Expose LedgerInfo only with testutils
