+++ soroban-env-host/src/host.rs
@@ -16,7 +16,6 @@ use soroban_env_common::xdr::{AccountId, Hash, PublicKey, ReadXdr, ThresholdInde
 
 use crate::budget::{Budget, CostType};
 use crate::events::{DebugError, DebugEvent, Events};
-use crate::ledger_info::LedgerInfo;
 use crate::storage::Storage;
 use crate::weak_host::WeakHost;
 
@@ -84,6 +83,14 @@ struct VmSlice {
     vm: Rc<Vm>,
     pos: u32,
     len: u32,
+}
+
+#[derive(Debug, Clone)]
+pub struct LedgerInfo {
+    pub protocol_version: u32,
+    pub sequence_number: u32,
+    pub timestamp: u64,
+    pub network_id: Vec<u8>,
 }
 
 #[derive(Clone, Default)]
