--- soroban-env-host/src/ledger_info.rs
@@ -1,7 +1,0 @@
-#[derive(Debug, Clone)]
-pub struct LedgerInfo {
-    pub protocol_version: u32,
-    pub sequence_number: u32,
-    pub timestamp: u64,
-    pub network_id: Vec<u8>,
-}
