+++ soroban-env-host/src/lib.rs
@@ -46,7 +46,8 @@ mod test;
 
 #[cfg(feature = "testutils")]
 pub use host::ContractFunctionSet;
+#[cfg(feature = "testutils")]
+pub use host::LedgerInfo;
 pub use host::{Host, HostError};
 pub use im_rc;
 pub use soroban_env_common::*;
-pub mod ledger_info;
